requirements.txt
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.10
chromadb==0.3.29
streamlit==1.31.0
python-dotenv==1.0.0
pypdf==3.17.1
openai>=1.6.1,<2.0.0
numpy==1.24.3
pandas==2.0.3
setuptools==69.0.3
wheel==0.42.0
onnxruntime==1.16.3 